+++ common/src/main/java/traben/entity_texture_features/config/screens/ETFConfigScreenSkinSettings.java
@@ -170,7 +170,8 @@ public class ETFConfigScreenSkinSettings extends ETFConfigScreen {
         context.drawCenteredTextWithShadow( textRenderer, ETFVersionDifferenceHandler.getTextFromTranslation("config." + ETFClientCommon.MOD_ID + ".player_skin_editor.button_desc.1"), (int) (width * 0.75), (int) (height * 0.4), 0xCCCCCC);
         context.drawCenteredTextWithShadow( textRenderer, ETFVersionDifferenceHandler.getTextFromTranslation("config." + ETFClientCommon.MOD_ID + ".player_skin_editor.button_desc.2"), (int) (width * 0.75), (int) (height * 0.45), 0xCCCCCC);
         if(!canLaunchTool)
-            context.drawCenteredTextWithShadow( textRenderer, ETFVersionDifferenceHandler.getTextFromTranslation("config." + ETFClientCommon.MOD_ID + ".player_skin_editor.button_desc.fail"), (int) (width * 0.75), (int) (height * 0.6), 0xCC5555);
+            context.drawCenteredTextWithShadow( textRenderer, ETFVersionDifferenceHandler.getTextFromTranslation("config." + ETFClientCommon.MOD_ID + ".player_skin_editor.button_desc.fail2"), (int) (width * 0.75), (int) (height * 0.6), 0xCC5555);
+        context.drawCenteredTextWithShadow( textRenderer, ETFVersionDifferenceHandler.getTextFromTranslation("config." + ETFClientCommon.MOD_ID + ".player_skin_editor.info"), (int) (width * 0.5), (int) (height * 0.8), 0xCCCCCC);
     }
 
 }